add extra tools, chrome fix shell square brackets

+++ installers/install-terraform.sh
@@ -1,17 +1,15 @@
 #!/bin/bash
 
-curl "https://releases.hashicorp.com/terraform/0.12.24/terraform_0.12.24_linux_amd64.zip" -o "/tmp/terraform.zip"
+if [ -z "$(which terraform)" ]; then
+    curl "https://releases.hashicorp.com/terraform/0.12.24/terraform_0.12.24_linux_amd64.zip" -o "/tmp/terraform.zip"
 
-unzip /tmp/terraform.zip
+    unzip -o /tmp/terraform.zip -d /tmp
 
-chmod +x /tmp/terraform
+    chmod +x /tmp/terraform
 
-sudo mv /tmp/terraform /usr/local/bin/
+    sudo mv /tmp/terraform /usr/local/bin/
 
-rm /tmp/terraform.zip
+    rm /tmp/terraform.zip
 
-echo "------------> terraform zip file removed succesfully"
-
-sudo chown root:root /usr/local/bin/terraform
-
-echo -ne "\n------------> terraform installed succesfully\n\n"
+    sudo chown root:root /usr/local/bin/terraform
+fi
